+++ src/apps/Map/components/WebglMap/WebglMap.unit.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import mapboxgl from 'mapbox-gl';
 import { render } from '~/utils/test-utils';
-import WebglMap from '.';
+import { WebglMap } from '.';
 
 describe('<WebglMap />', () => {
   it('renders', () => {
